GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - Adding the log library (refs/gitbutler/Adding-the-log-library)
   branch head: 666afbdd7559a7b1778eefaea9640964b3f3ca8c
   - src/main.cpp
 - feat--adding-the-AS726X-library (refs/gitbutler/feat--adding-the-AS726X-library)
   branch head: f5188958842e3ad7404a85b375e1fa559bbfb59e
For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch